+++ src/behaviors/index.js
@@ -8,8 +8,11 @@ import { assign, pick, values } from 'lodash';
 
 import { id_from_title } from './id_from_title/id_from_title';
 
+const { config } = require(`${process.cwd()}/config`);
+
 const behaviors = {
   id_from_title,
+  ...('behaviors' in config ? config.behaviors : {}),
 };
 
 /**