+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add support for multiple profiles @robgietema
 - Add yeoman generator to create a Nick project @robgietema
 - Add vocabularies based on profiles @robgietema
+- Make behaviors and vocabularies extendable @robgietema
 
 ### Bugfix